2 byte pin index #4097 (#4102)
* 2 byte pin index #4097

* 2 byte pin index #4097

good time to defragment configuration

* 2 byte pin index #4097

new fancy way

* 2 byte pin index #4097

smart comment

+++ firmware/config/boards/subaru_eg33/rusefi_hw_enums.h
@@ -20,7 +20,7 @@ typedef enum __attribute__ ((__packed__)) {
 /**
  * Hardware pin. This enum is platform-specific.
  */
-enum class Gpio : uint8_t {
+enum class Gpio : uint16_t {
 	Unassigned = 0,
 	Invalid = 1,
